Separated Dto objects contained in MetarDto to separate class files.
