+++ app/src/main/java/io/mcnulty/avwx/data/remote/dto/metar/AltimeterDto.kt
@@ -1,0 +1,7 @@
+package io.mcnulty.avwx.data.remote.dto.metar
+
+data class AltimeterDto(
+    val repr: String,
+    val spoken: String,
+    val value: Number
+)
