+++ app/src/main/java/io/mcnulty/avwx/data/remote/dto/metar/CloudDto.kt
@@ -1,0 +1,8 @@
+package io.mcnulty.avwx.data.remote.dto.metar
+
+data class CloudDto(
+    val altitude: Int,
+    val modifier: String,
+    val repr: String,
+    val type: String
+)
