+++ app/src/main/java/io/mcnulty/avwx/data/remote/dto/metar/DewpointDto.kt
@@ -1,0 +1,7 @@
+package io.mcnulty.avwx.data.remote.dto.metar
+
+data class DewpointDto(
+    val repr: String,
+    val spoken: String,
+    val value: Int
+)
