+++ app/src/main/java/io/mcnulty/avwx/data/remote/dto/metar/MetaData.kt
@@ -1,0 +1,11 @@
+package io.mcnulty.avwx.data.remote.dto.metar
+
+import com.google.gson.annotations.SerializedName
+
+data class MetaData(
+    @SerializedName("cache_timestamp")
+    val cacheTimestamp: String,
+    @SerializedName("stations_updated")
+    val stationsUpdated: String,
+    val timestamp: String
+)
