+++ app/src/main/java/io/mcnulty/avwx/data/remote/dto/metar/TimeDto.kt
@@ -1,0 +1,6 @@
+package io.mcnulty.avwx.data.remote.dto.metar
+
+data class TimeDto(
+    val dt: String,
+    val repr: String
+)
