+++ app/src/main/java/io/mcnulty/avwx/data/remote/dto/metar/VisibilityDto.kt
@@ -1,0 +1,7 @@
+package io.mcnulty.avwx.data.remote.dto.metar
+
+data class VisibilityDto(
+    val repr: String,
+    val spoken: String,
+    val value: Int
+)
