+++ app/src/main/java/io/mcnulty/avwx/data/remote/dto/metar/WeatherCodeDto.kt
@@ -1,0 +1,6 @@
+package io.mcnulty.avwx.data.remote.dto.metar
+
+data class WeatherCodeDto(
+    val repr: String,
+    val value: String
+)
