+++ app/src/main/java/io/mcnulty/avwx/data/remote/dto/metar/WindSpeed.kt
@@ -1,0 +1,7 @@
+package io.mcnulty.avwx.data.remote.dto.metar
+
+data class WindSpeed(
+    val repr: String,
+    val spoken: String,
+    val value: Int
+)
