+++ app/src/main/java/io/mcnulty/avwx/domain/use_case/parse/TemperatureParser.kt
@@ -1,6 +1,6 @@
 package io.mcnulty.avwx.domain.use_case.parse
 
-import io.mcnulty.avwx.data.remote.dto.MetarDto
+import io.mcnulty.avwx.data.remote.dto.metar.MetarDto
 import io.mcnulty.avwx.domain.model.metar.fields.Temperature
 import io.mcnulty.avwx.domain.model.metar.measurement.TemperatureUnits
 
@@ -21,6 +21,6 @@ internal object TemperatureParser {
     ): Temperature {
         return dto.remarksInfo.temperatureDecimal?.let { tempDecimal ->
             Temperature(tempDecimal.value, units)
-        } ?: Temperature(dto.temperature.value.toDouble(), units)
+        } ?: Temperature(dto.temperatureDto.value.toDouble(), units)
     }
 }
